app/requirements.txt
asgiref==3.2.10
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.7.0rc1
Django==3.1.1
django-contrib-comments==2.0.0
django-crispy-forms==1.10.0
django-filter==2.4.0
git+https://github.com/django-fluent/django-fluent-comments.git#egg=django_fluent_comments
django-nested-admin==3.3.3
django-nested-formset==0.1.4
django-recaptcha==2.0.6
django-tag-parser==3.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-akismet==0.4.2
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.3.1
urllib3==1.26.2
